+++ demos/37-conways-game-of-life/detectDarkMode.js
@@ -1,0 +1,7 @@
+function detectDarkMode() {
+  var hours = new Date().getHours();
+  var isDayTime = hours > 6 && hours < 20;
+  return !isDayTime;
+}
+
+globalThis.isDarkMode = detectDarkMode();